Add AimingControl. Use this class to implement rotate - and - aim behaviour.

+++ Defence1/Assets/Scripts/MapGenerator.cs
@@ -5,19 +5,19 @@ using System;
 public class MapGenerator {
 
 	public readonly float oreNum;
-	public readonly float minOreValue = 50.0f;
-	public readonly float maxOreValue = 1000.0f;
+	public readonly float minOreValue = 50;
+	public readonly float maxOreValue = 1000;
 
 	public MapGenerator (float oreNum){
 		this.oreNum = oreNum;
 	}
 
-	int randomOreNum(){
+	float randomOreNum(){
 		var x = UnityEngine.Random.value;
-		return (int)(minOreValue + x * x * (maxOreValue - minOreValue));
+		return (minOreValue + x * x * (maxOreValue - minOreValue));
 	}
 
-	public void generateOres(Action<Vector2,int> genFunc,Func<Vector2> randomPosInScene){
+	public void generateOres(Action<Vector2,float> genFunc,Func<Vector2> randomPosInScene){
 
 		for (int i = 0; i < oreNum; i++) {
 			var pos = randomPosInScene ();
